Control updates

+++ extensions/edu/berkeley/guir/prefusex/controls/FocusControl.java
@@ -157,10 +157,12 @@ public class FocusControl extends ControlAdapter {
             if ( ccount == 0 ) {
         		curFocus = item.getEntity();
         		ItemRegistry registry = item.getItemRegistry();
-        		FocusManager fm = registry.getFocusManager();
-        		FocusSet fs = fm.getFocusSet(focusSetKey);
-        		fs.set(item.getEntity());
-        		registry.touch(item.getItemClass());
+        		synchronized ( registry ) {
+	        		FocusManager fm = registry.getFocusManager();
+	        		FocusSet fs = fm.getFocusSet(focusSetKey);
+	        		fs.set(item.getEntity());
+	        		registry.touch(item.getItemClass());
+        		}
         		runActivity();
             }
         }
@@ -173,10 +175,14 @@ public class FocusControl extends ControlAdapter {
             if ( ccount == 0 ) {
             	curFocus = null;
                 ItemRegistry registry = item.getItemRegistry();
-                FocusManager fm = registry.getFocusManager();
-                FocusSet fs = fm.getFocusSet(focusSetKey);
-                fs.remove(item.getEntity());
-                registry.touch(item.getItemClass());
+                if ( registry != null ) {
+                    synchronized ( registry ) {
+	                    FocusManager fm = registry.getFocusManager();
+	                    FocusSet fs = fm.getFocusSet(focusSetKey);
+	                    fs.remove(item.getEntity());
+	                    registry.touch(item.getItemClass());
+                    }
+                }
                 runActivity();
             }
         }
@@ -189,12 +195,22 @@ public class FocusControl extends ControlAdapter {
         {
         	Entity focus = item.getEntity();
         	if ( focus != curFocus ) {
-        		curFocus = focus;
 	            ItemRegistry registry = item.getItemRegistry();
-	            FocusManager fm = registry.getFocusManager();
-	            FocusSet fs = fm.getFocusSet(focusSetKey);
-	            fs.set(item.getEntity());
-	            registry.touch(item.getItemClass());
+	            synchronized ( registry ) {
+		            FocusManager fm = registry.getFocusManager();
+		            FocusSet fs = fm.getFocusSet(focusSetKey);
+		            
+		            boolean ctrl = e.isControlDown();
+		            if ( !ctrl ) {
+		                curFocus = focus;
+		                fs.set(focus);
+		            } else if ( fs.contains(focus) ) {
+		                fs.remove(focus);
+		            } else {
+		                fs.add(focus);
+		            }
+		            registry.touch(item.getItemClass());
+	            }
 	            runActivity();
         	}
         }
